requirements.txt
imgurpython
python-telegram-bot
